requirements.txt
streamlit
transformers
torch
torchaudio
librosa
opencv-python-headless
Pillow
scikit-learn
moviepy
ffmpeg-python
